Security login logout

+++ spring-boot-accounting-book/src/main/java/com/xintongthecoder/accountingbook/controller/AccountBookController.java
@@ -6,9 +6,9 @@ import org.springframework.data.web.PagedResourcesAssembler;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.MediaTypes;
 import org.springframework.hateoas.PagedModel;
+import org.springframework.hateoas.server.ExposesResourceFor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,36 +17,41 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import com.xintongthecoder.accountingbook.dao.AccountBookRepository;
+import com.xintongthecoder.accountingbook.dao.AccountRepository;
 import com.xintongthecoder.accountingbook.entity.Account;
 import com.xintongthecoder.accountingbook.entity.AccountBook;
+import com.xintongthecoder.accountingbook.errorHandler.AccountAccessDeniedException;
 import com.xintongthecoder.accountingbook.errorHandler.AccountBookNotFoundException;
 import com.xintongthecoder.accountingbook.modelAssembler.AccountBookModelAssembler;
 
 
-@CrossOrigin("http://localhost:4200")
 @RestController
-@RequestMapping("api/books")
+@RequestMapping("api/accounts")
+@ExposesResourceFor(AccountBook.class)
 public class AccountBookController {
 
         private final AccountBookRepository accountBookRepository;
         private final AccountBookModelAssembler accountBookModelAssembler;
         private final PagedResourcesAssembler<AccountBook> pagedResourcesAssembler;
+        private final AccountRepository accountRepository;
 
         public AccountBookController(AccountBookRepository accountBookRepository,
                         AccountBookModelAssembler accountBookModelAssembler,
-                        PagedResourcesAssembler<AccountBook> pagedResourcesAssembler) {
+                        PagedResourcesAssembler<AccountBook> pagedResourcesAssembler,
+                        AccountRepository accountRepository) {
                 this.accountBookRepository = accountBookRepository;
                 this.accountBookModelAssembler = accountBookModelAssembler;
                 this.pagedResourcesAssembler = pagedResourcesAssembler;
+                this.accountRepository = accountRepository;
         }
 
-        @GetMapping(value = "/{bookId}", produces = {"application/hal+json"})
-        public ResponseEntity<PagedModel<EntityModel<AccountBook>>> one(@PathVariable Long bookId,
-                        @RequestParam(value = "page", defaultValue = "0") int page,
-                        @RequestParam(value = "size", defaultValue = "10") int size) {
+        @GetMapping(value = "{email}/books/{bookId}", produces = {"application/hal+json"})
+        public ResponseEntity<PagedModel<EntityModel<AccountBook>>> one(
+                        @PathVariable("email") String email, @PathVariable Long bookId,
+                        @RequestParam(value = "page", defaultValue = "0") Integer page,
+                        @RequestParam(value = "size", defaultValue = "10") Integer size) {
                 Page<AccountBook> pagedBook =
                                 accountBookRepository.findById(bookId, PageRequest.of(page, size));
                 return ResponseEntity.ok().contentType(MediaTypes.HAL_JSON)
@@ -54,38 +59,55 @@ public class AccountBookController {
                                                 accountBookModelAssembler));
         }
 
-        @GetMapping(value = "", produces = {"application/hal+json"})
+        private boolean requestFromAuthorizedAccount(String email, Long bookId) {
+                if (accountBookRepository.getReferenceById(bookId) == null) {
+                        throw new AccountBookNotFoundException(bookId);
+                }
+                return accountBookRepository.findById(bookId).get().getAccount().getEmail()
+                                .equals(email);
+        }
+
+        @GetMapping(value = "/{email}/books", produces = {"application/hal+json"})
         public ResponseEntity<PagedModel<EntityModel<AccountBook>>> all(
-                        @RequestParam(value = "page", defaultValue = "0") int page,
-                        @RequestParam(value = "size", defaultValue = "10") int size) {
-                Page<AccountBook> pagedBooks =
-                                accountBookRepository.findAll(PageRequest.of(page, size));
+                        @PathVariable("email") String email,
+                        @RequestParam(value = "page", defaultValue = "0") Integer page,
+                        @RequestParam(value = "size", defaultValue = "10") Integer size) {
+                Account account = accountRepository.findByEmail(email);
+                Page<AccountBook> pagedBooks = accountBookRepository.findAllByAccount(account,
+                                PageRequest.of(page, size));
                 return ResponseEntity.ok().contentType(MediaTypes.HAL_JSON)
                                 .body(pagedResourcesAssembler.toModel(pagedBooks,
                                                 accountBookModelAssembler));
         }
 
-        @PostMapping(value = "", produces = {"application/hal+json"})
-        public ResponseEntity<AccountBook> addBook(@RequestBody AccountBook bookFromRequest) {
-                // TODO: Set accountId
+        @PostMapping(value = "/{email}/books", produces = {"application/hal+json"})
+        public ResponseEntity<AccountBook> addBook(@PathVariable("email") String email,
+                        @RequestBody AccountBook bookFromRequest) {
+                Account account = accountRepository.findByEmail(email);
+                bookFromRequest.setAccount(account);
                 AccountBook newBook = this.accountBookRepository.save(bookFromRequest);
                 return new ResponseEntity<>(newBook, HttpStatus.CREATED);
         }
 
-        @PutMapping(value = "/{bookId}", produces = {"application/hal+json"})
-        public ResponseEntity<AccountBook> editBook(@PathVariable("bookId") Long bookId,
+        @PutMapping(value = "{email}/books/{bookId}", produces = {"application/hal+json"})
+        public ResponseEntity<AccountBook> editBook(@PathVariable("email") String email,
+                        @PathVariable("bookId") Long bookId,
                         @RequestBody AccountBook bookFromRequest) {
-                if (accountBookRepository.getReferenceById(bookId) == null) {
-                        throw new AccountBookNotFoundException(bookId);
+                if (!requestFromAuthorizedAccount(email, bookId)) {
+                        throw new AccountAccessDeniedException("book");
                 }
-                bookFromRequest.setAccount(
-                                accountBookRepository.getReferenceById(bookId).getAccount());
+                Account account = accountRepository.findByEmail(email);
+                bookFromRequest.setAccount(account);
                 AccountBook updatedBook = accountBookRepository.save(bookFromRequest);
                 return new ResponseEntity<AccountBook>(updatedBook, HttpStatus.OK);
         }
 
-        @DeleteMapping(value = "/{bookId}", produces = {"application/hal+json"})
-        public ResponseEntity<HttpStatus> deleteBook(@PathVariable("bookId") Long bookId) {
+        @DeleteMapping(value = "{email}/books/{bookId}", produces = {"application/hal+json"})
+        public ResponseEntity<HttpStatus> deleteBook(@PathVariable("email") String email,
+                        @PathVariable("bookId") Long bookId) {
+                if (!requestFromAuthorizedAccount(email, bookId)) {
+                        throw new AccountAccessDeniedException("book");
+                }
                 accountBookRepository.deleteById(bookId);
                 return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
